Add files via upload
There are two folders, and each of them stands for a method of realizing the decoder. -->See Readme-Refernces for details
And in each file:
The ".m" is the m file including both encoder and decoder of BCH(63,56), and the look up table it use is saved in the ".xls" file.
The .v files are the Verilog HDL files.